+++ src/GraphicObjects/Grid.cs
@@ -32,13 +32,13 @@ namespace Crow
 		public override T addChild<T> (T child)
 		{
 			T tmp = base.addChild (child);
-			this.RegisterForLayouting ((int)LayoutingType.PositionChildren);
+			this.RegisterForLayouting (LayoutingType.PositionChildren);
 			return tmp;
 		}
 		public override void removeChild (GraphicObject child)
 		{
 			base.removeChild (child);
-			this.RegisterForLayouting ((int)LayoutingType.PositionChildren);
+			this.RegisterForLayouting (LayoutingType.PositionChildren);
 		}
 
 		#region Public Properties
@@ -59,7 +59,7 @@ namespace Crow
 				_columnCount = value; 
 
 				NotifyValueChanged ("ColumnCount", ColumnCount);
-				this.RegisterForLayouting ((int)LayoutingType.PositionChildren);
+				this.RegisterForLayouting (LayoutingType.PositionChildren);
 			}
         }
 		[XmlAttributeAttribute()][DefaultValue(1)]
@@ -73,7 +73,7 @@ namespace Crow
 				_rowCount = value; 
 
 				NotifyValueChanged ("RowCount", RowCount);
-				this.RegisterForLayouting ((int)LayoutingType.PositionChildren);
+				this.RegisterForLayouting (LayoutingType.PositionChildren);
 			}
 		}
 		public virtual int CaseWidth {
